refactor: unify signal package exports
