+++ tests/test_signal_generator_misc.py
@@ -4,7 +4,7 @@ from collections import deque, OrderedDict
 
 from quant_trade.robust_signal_generator import RobustSignalGenerator
 from quant_trade.market_phase import get_market_phase
-from quant_trade.signal.factor_scorer import FactorScorerImpl
+from quant_trade.signal import FactorScorerImpl
 
 
 def make_rsg():
